processes plugin: support of matching by user
The configuration of "ProcessMatch" was extended by an
optional third parameter. It is a regular expression for the
user name under which the process runs.

The new configuration parameter 'ProcessMatchByUser'
automatically creates a separate group for all users existing
on the system. The name of the group corresponds to the
user name.